Add: helper to keep strong references to created tasks
Python's asyncio.create_task() keeps an internal weak reference.
In result, the GC can remove the task if the returning task is
not tracked.

But it is much more convenient to use that function as a
fire&forget, and to not always have to track it.

So, instead of having to add code in other places to send these
tasks to a background-queue, take over the whole factory and
create a strong reference for every task created.